Client basic implementation, refactoring
