+++ internal/infra/transport/client/client.go
@@ -1,43 +1,66 @@
 package client
 
 import (
+	"bufio"
 	"context"
 	"errors"
 	"fmt"
 	"net"
+	"slices"
 
 	"github.com/bullgare/pow-ddos-protection/internal/infra/protocol/common"
-	"github.com/bullgare/pow-ddos-protection/internal/infra/transport/connection"
+	"github.com/bullgare/pow-ddos-protection/internal/infra/transport"
 )
 
 func New(
 	address string,
-	connHandler *connection.Connection,
 ) (*Client, error) {
 	if address == "" {
 		return nil, errors.New("address is required")
 	}
-	if connHandler == nil {
-		return nil, errors.New("connection handler is required")
-	}
 
 	return &Client{
-		address:     address,
-		connHandler: connHandler,
+		address: address,
 	}, nil
 }
 
 type Client struct {
-	address     string
-	connHandler *connection.Connection
+	address string
 }
 
-func (c Client) SendRequest(ctx context.Context, req common.Request) (common.Response, error) {
+func (c *Client) SendRequest(ctx context.Context, req common.Request) (common.Response, error) {
 	conn, err := net.Dial("tcp", c.address)
 	if err != nil {
 		return common.Response{}, fmt.Errorf("connecting to %q: %w", c.address, err)
 	}
 	defer func() { _ = conn.Close() }()
 
-	return c.connHandler.SendRequest(ctx, conn, req)
+	r := bufio.NewReader(conn)
+	w := bufio.NewWriter(conn)
+
+	msg := common.Message{
+		Version: common.MessageVersionV1,
+		Type:    req.Type,
+		Payload: slices.Clone(req.Payload),
+	}
+
+	err = transport.SendMessage(w, msg)
+	if err != nil {
+		return common.Response{}, fmt.Errorf("sending request %v: %w", req, err)
+	}
+
+	rawResp, err := r.ReadString(transport.MessageDelimiter)
+	if err != nil {
+		return common.Response{}, fmt.Errorf("reading response: %w", err)
+	}
+
+	msg, err = transport.ParseRawMessage(rawResp)
+	if err != nil {
+		return common.Response{}, fmt.Errorf("parsing request: %w", err)
+	}
+
+	return common.Response{
+		Type:    msg.Type,
+		Payload: slices.Clone(msg.Payload),
+	}, nil
 }
