+++ internal/infra/transport/connection.go
@@ -1,34 +1,16 @@
-package connection
+package transport
 
 import (
 	"bufio"
-	"errors"
 	"fmt"
 	"strings"
 
 	"github.com/bullgare/pow-ddos-protection/internal/infra/protocol/common"
 )
 
-func New(
-	onError func(error),
-) (*Connection, error) {
-	if onError == nil {
-		return nil, errors.New("onError is required")
-	}
+var MessageDelimiter = []byte(common.Terminator)[0]
 
-	return &Connection{
-		onError:   onError,
-		delimiter: []byte(common.Terminator)[0],
-	}, nil
-}
-
-// Connection knows how to send and receive messages over a
-type Connection struct {
-	onError   func(error)
-	delimiter byte
-}
-
-func (c *Connection) sendMessage(w *bufio.Writer, msg common.Message) error {
+func SendMessage(w *bufio.Writer, msg common.Message) error {
 	raw := generateRawMessage(msg)
 
 	_, err := w.WriteString(raw)
@@ -44,7 +26,7 @@ func (c *Connection) sendMessage(w *bufio.Writer, msg common.Message) error {
 	return nil
 }
 
-func (c *Connection) parseRawMessage(raw string) (common.Message, error) {
+func ParseRawMessage(raw string) (common.Message, error) {
 	raw = strings.TrimSpace(raw)
 	chunks := strings.Split(raw, common.Separator)
 	if len(chunks) < 3 {
